pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-crew-blog-scraper"
version = "0.1.0"
description = "Blog scraping pipeline for Context Crew capstone project"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Context Crew Team"}
]
keywords = ["rag", "llm", "blog-scraper", "contextual-ai", "web-scraping"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "contextual-client>=0.10.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.2",
    "markdownify>=0.11.6",
    "feedparser>=6.0.10",
    "structlog>=23.2.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "responses>=0.24.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
]

[project.scripts]
scrape-blogs = "scripts.run_scraper:main"
upload-to-contextual = "scripts.upload_to_contextual:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "venv", "build", "dist"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
